moved filter out of builds table into dashboard page.
